third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-javascript-3.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/debugger-test.js"></script>
<script src="../../inspector/sources-test.js"></script>
<script src="debugger/resources/obfuscated.js"></script>

<script>

function test() {
  var testJSFormatter = SourcesTestRunner.testPrettyPrint.bind(InspectorTest, 'text/javascript');

  TestRunner.runTestSuite([
    function ifStatements(next) {
      var mappingQueries = ['===', '!==', 'non-eq'];
      testJSFormatter(
          'if(a<b)log(a);else log(b);if(a<b){log(a)}else{log(b);}if(a===b)log(\'equals\');if(a!==b){log(\'non-eq\');}',
          mappingQueries, next);
    },

    function arrayLiteralFormatting(next) {
      var mappingQueries = ['3', '2', '1', '0'];
      testJSFormatter('var arr=[3,2,1,0]', mappingQueries, next);
    },

    function ifFormatting(next) {
      var mappingQueries = ['&&', 'print(a)'];
      testJSFormatter('if(a>b&&b>c){print(a);print(b);}', mappingQueries, next);
    },

    function ternarOperatorFormatting(next) {
      var mappingQueries = ['?', ':'];
      testJSFormatter('a>b?a:b', mappingQueries, next);
    },

    function labeledStatementFormatting(next) {
      var mappingQueries = ['break', 'continue', 'while'];
      testJSFormatter('firstLoop:while(true){break firstLoop;continue firstLoop;}', mappingQueries, next);
    },
  ]);
}

</script>

</head>

<body onload="runTest()">
<p>Verifies JavaScript pretty-printing functionality.</p>
</body>
</html>
