third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-javascript-3-expected.txt

Verifies JavaScript pretty-printing functionality.


Running: ifStatements
====== 8< ------
if (a < b)
    log(a);
else
    log(b);
if (a < b) {
    log(a)
} else {
    log(b);
}
if (a === b)
    log('equals');
if (a !== b) {
    log('non-eq');
}

------ >8 ======
Correct mapping for <===>
Correct mapping for <!==>
Correct mapping for <non-eq>

Running: arrayLiteralFormatting
====== 8< ------
var arr = [3, 2, 1, 0]

------ >8 ======
Correct mapping for <3>
Correct mapping for <2>
Correct mapping for <1>
Correct mapping for <0>

Running: ifFormatting
====== 8< ------
if (a > b && b > c) {
    print(a);
    print(b);
}

------ >8 ======
Correct mapping for <&&>
Correct mapping for <print(a)>

Running: ternarOperatorFormatting
====== 8< ------
a > b ? a : b

------ >8 ======
Correct mapping for <?>
Correct mapping for <:>

Running: labeledStatementFormatting
====== 8< ------
firstLoop: while (true) {
    break firstLoop;
    continue firstLoop;
}

------ >8 ======
Correct mapping for <break>
Correct mapping for <continue>
Correct mapping for <while>
